fix(csharp): use Int32 for BUFFER_LENGTH and accept both Int8/Int32 in tests
BUFFER_LENGTH uses Int32 in the SEA schema factory. Thrift server returns
Int8. Tests now accept both types so they pass regardless of protocol.

+++ csharp/src/DatabricksStatement.cs
@@ -827,7 +827,7 @@ namespace AdbcDrivers.Databricks
             var dataTypeBuilder = new Int32Array.Builder();
             var typeNameBuilder = new StringArray.Builder();
             var columnSizeBuilder = new Int32Array.Builder();
-            var bufferLengthBuilder = new Int8Array.Builder();
+            var bufferLengthBuilder = new Int32Array.Builder();
             var decimalDigitsBuilder = new Int32Array.Builder();
             var numPrecRadixBuilder = new Int32Array.Builder();
             var nullableBuilder = new Int32Array.Builder();
